requirement.txt
alembic==1.6.4
appdirs==1.4.4
black==21.6b0
boto3==1.17.101
botocore==1.20.101
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colored==1.4.2
configobj==5.0.6
distlib==0.3.1
Faker==8.8.2
filelock==3.0.12
Flask==1.1.4
flask-alchemydumps==0.0.11
Flask-Migrate==2.5.3
Flask-Script==1.0.0
Flask-SQLAlchemy==2.4.3
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-dataclass==8.4.1
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumbers==8.12.24
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-editor==1.0.4
regex==2021.4.4
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
slack-cleaner==0.7.4
slacker==0.14.0
SQLAlchemy==1.4.13
supervisor==4.2.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
virtualenv==20.4.7
Werkzeug==1.0.1
